Refactor: TypedDicts, decomposition, trend gate, coverage gate, docs
- TypedDicts adopted throughout the codebase (AuditResult, LizardSummary,
  SccRecord, GitInfo, DiffSummary) replacing bare dict[str, Any] signatures
- export_outputs decomposed into _enrich_results, _export_json, _export_csv,
  _export_markdown (CCN 14 -> 3)
- _init_db made declarative with _MIGRATIONS list (119 -> 40 NLOC)
- _save_results decomposed into _upsert_metrics_row, _log_metrics_history,
  _save_file_metrics (79 -> 15 NLOC)
- Trend gate now queries metrics_history (append-only) instead of metrics
- Coverage gate in CI (60%)
- README documents runs_count and upsert
- Fixed hardcoded file_metric warnings
- Fixed migration loop bug (version = 7 outside if block)

+++ scopio/types.py
@@ -67,11 +67,11 @@ class AuditResult(TypedDict, total=False):
 
 
 class DiffDelta(TypedDict, total=False):
-    loc: int
+    loc: float
     loc_trend: float | None
     ccn: float
     ccn_trend: float | None
-    warnings: int
+    warnings: float
 
 
 class DiffSnapshot(TypedDict, total=False):
